consolidate remaining types into shared package
- Add BitcoinConfig, JdcConfig, TranslatorConfig, SetupData to shared
- Add all log diagnostic types (LogContainerRole, ContainerLogLine, etc.) to shared
- Update server/src/types.ts to import from shared
- Update server/src/logs/types.ts to import from shared
- Update src/components/setup/types.ts to import from shared
- Update src/types/log-diagnostics.ts to import from shared
