+++ server/src/types.ts
@@ -1,43 +1,6 @@
-import type { MiningMode, SetupMode, OperatingSystem, BitcoinCoreVersion, BitcoinNetwork, HealthStatus } from '@sv2-ui/shared';
+import type { MiningMode, SetupMode, OperatingSystem, BitcoinCoreVersion, BitcoinNetwork, HealthStatus, PoolConfig, BitcoinConfig, JdcConfig, TranslatorConfig, SetupData } from '@sv2-ui/shared';
 
-export interface PoolConfig {
-  name: string;
-  address: string;
-  port: number;
-  authority_public_key: string;
-}
-
-export interface BitcoinConfig {
-  core_version: BitcoinCoreVersion | null;
-  network: BitcoinNetwork;
-  os: OperatingSystem;
-  customDataDir: string;
-  socket_path: string;
-}
-
-export interface JdcConfig {
-  user_identity: string;
-  jdc_signature: string;
-  coinbase_reward_address: string;
-}
-
-export interface TranslatorConfig {
-  user_identity: string;
-  enable_vardiff: boolean;
-  aggregate_channels: boolean;
-  min_hashrate: number;
-  shares_per_minute: number;
-  downstream_extranonce2_size: number;
-}
-
-export interface SetupData {
-  miningMode: MiningMode;
-  mode: SetupMode;
-  pool: PoolConfig | null;
-  bitcoin: BitcoinConfig | null;
-  jdc: JdcConfig | null;
-  translator: TranslatorConfig;
-}
+export type { PoolConfig, BitcoinConfig, JdcConfig, TranslatorConfig, SetupData };
 
 export interface ContainerStatus {
   id: string;
